Nest existing tests under onError folders
