+++ dev-packages/browser-integration-tests/suites/integrations/featureFlags/openfeature/onError/basic/test.ts
@@ -1,7 +1,7 @@
 import { expect } from '@playwright/test';
-import { sentryTest } from '../../../../../utils/fixtures';
-import { envelopeRequestParser, shouldSkipFeatureFlagsTest, waitForErrorRequest } from '../../../../../utils/helpers';
-import { FLAG_BUFFER_SIZE } from '../../constants';
+import { sentryTest } from '../../../../../../utils/fixtures';
+import { envelopeRequestParser, shouldSkipFeatureFlagsTest, waitForErrorRequest } from '../../../../../../utils/helpers';
+import { FLAG_BUFFER_SIZE } from '../../../constants';
 
 sentryTest('Basic test with eviction, update, and no async tasks', async ({ getLocalTestUrl, page }) => {
   if (shouldSkipFeatureFlagsTest()) {
@@ -20,13 +20,12 @@ sentryTest('Basic test with eviction, update, and no async tasks', async ({ getL
   await page.goto(url);
 
   await page.evaluate(bufferSize => {
-    const ldClient = (window as any).initializeLD();
+    const client = (window as any).initialize();
     for (let i = 1; i <= bufferSize; i++) {
-      ldClient.variation(`feat${i}`, false);
+      client.getBooleanValue(`feat${i}`, false);
     }
-    ldClient.variation(`feat${bufferSize + 1}`, true); // eviction
-    ldClient.variation('feat3', true); // update
-    return true;
+    client.getBooleanValue(`feat${bufferSize + 1}`, true); // eviction
+    client.getBooleanValue('feat3', true); // update
   }, FLAG_BUFFER_SIZE);
 
   const reqPromise = waitForErrorRequest(page);
