+++ dev-packages/browser-integration-tests/suites/integrations/featureFlags/openfeature/onError/errorHook/init.js
