+++ dev-packages/browser-integration-tests/suites/integrations/featureFlags/openfeature/onError/errorHook/test.ts
@@ -1,7 +1,7 @@
 import { expect } from '@playwright/test';
-import { sentryTest } from '../../../../../utils/fixtures';
-import { envelopeRequestParser, shouldSkipFeatureFlagsTest, waitForErrorRequest } from '../../../../../utils/helpers';
-import { FLAG_BUFFER_SIZE } from '../../constants';
+import { sentryTest } from '../../../../../../utils/fixtures';
+import { envelopeRequestParser, shouldSkipFeatureFlagsTest, waitForErrorRequest } from '../../../../../../utils/helpers';
+import { FLAG_BUFFER_SIZE } from '../../../constants';
 
 sentryTest('Flag evaluation error hook', async ({ getLocalTestUrl, page }) => {
   if (shouldSkipFeatureFlagsTest()) {
