+++ dev-packages/browser-integration-tests/suites/integrations/featureFlags/openfeature/onError/init.js
