+++ dev-packages/browser-integration-tests/suites/integrations/featureFlags/statsig/onError/init.js
